[FIXED STAPLER-8] In looking at HUDSON-2670 I found that stapler did not recognize the difference between abc/def and abc%2Fdef in a url.. it would always try to map the request to "abc" instead of "abc/def" as a single entity.
Changed Stapler.getServletPath to use getRequestURI instead of getServletPath
because the former gives a raw value (not decoded).  Now when TokenList splits
on "/" it will use the right boundaries, not a %2F.  TokenList decodes the
individual elements, except in assembleRestOfPath which is used in generating
new urls so we want rawTokens here.  Also updated AncestorImpl to use rawTokens
for url generation.. was then able to remove the new URI(...).toString() part..
don't need to re-encode when using rawTokens that were not decoded.



git-svn-id: https://svn.dev.java.net/svn/stapler/trunk/stapler@1102 35881493-e564-0410-9e88-c8203b5c1b61